Add retries to write response calls to proxy
The goal is to avoid failing the end user request in case of a transient
network connectivity error when forwarding a response to the proxy,
assuming it succeeds on one of the next attempts. The retries are only
applied to errors that occur early on, before the proxy starts reading
the response body.